+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
  - Remove unused `pristineValue` property in `Field`.
  - Add TypeScript type definitions.
+ - Add `moveUp` and `moveDown` to `ListForm`.
 
 ## v1.2.0
  - Support identification of hierarchical touched state.